Add ssh-menu
ssh-menu is an ncurses application for managing ssh connections.
It provides a menu for selecting different hosts and establishing
a connection to the selected host based on the information
provided in ~/.ssh/menu.

ssh-menu is useful in combination with xterm and keybindings for
the window manager. I am using it with xmonad and the additional
binding

((mod1Mask .|. shiftMask, xK_backslash),
    spawn "xterm -e ssh-menu")