Add white-ice-cron script
white-ice-cron can be used in a cron job to monitor white-ice and
restart it if necessary. The script assumes that white-ice.py is
installed under /root/bin/white-ice with executable permissions.
Currently white-ice does not write a pid file. This would clearly
make white-ice-cron simpler and more elegant but for now this
solution seems accetable.