white-ice: Do not store the rule number
The rule numbers are changing all the time when blocking new
clients so it really makes no sense to store it as part of
the client. We rather should retrieve it just-in-time in the
unblock() function. Since we generally do not support unblocking
of clients this has not been an issue before.